refactor: Move action logic to App class
